requirements.txt
beautifulsoup4 (4.6.0)
certifi (2017.11.5)
chardet (3.0.4)
flake8 (3.5.0)
idna (2.6)
mccabe (0.6.1)
pip (9.0.1)
pycodestyle (2.3.1)
pyflakes (1.6.0)
python-Levenshtein (0.12.0)
requests (2.18.4)
setuptools (38.4.0)
urllib3 (1.22)
wheel (0.30.0)
